+++ src/main/java/com/sammccreery/cherry/Cherry.java
@@ -1,12 +1,17 @@
 package com.sammccreery.cherry;
 
 import com.sammccreery.cherry.event.Events;
+import com.sammccreery.cherry.registry.CherryBlocks;
+import com.sammccreery.cherry.registry.CherryGeneration;
 import com.sammccreery.cherry.registry.CherryItems;
+import com.sammccreery.cherry.registry.Registry;
 
+import cpw.mods.fml.common.IWorldGenerator;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.Mod.EventHandler;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.common.MinecraftForge;
+import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 
 @Mod(modid=Cherry.MODID, useMetadata=true)
 public class Cherry {
@@ -14,7 +19,10 @@ public class Cherry {
 
 	@EventHandler
 	public void init(FMLInitializationEvent e) {
-		CherryItems.init();
-		MinecraftForge.EVENT_BUS.register(new Events());
+		Registry.addRegister(Item.class, new CherryItems());
+		Registry.addRegister(Block.class, new CherryBlocks());
+		Registry.addRegister(Object.class, new Events());
+		Registry.addRegister(IWorldGenerator.class, new CherryGeneration());
+		Registry.initAll();
 	}
 }