+++ src/main/java/com/sammccreery/cherry/generation/WorldGenOres.java
@@ -1,0 +1,15 @@
+package com.sammccreery.cherry.generation;
+
+import java.util.Random;
+
+import cpw.mods.fml.common.IWorldGenerator;
+import net.minecraft.world.World;
+import net.minecraft.world.chunk.IChunkProvider;
+
+public class WorldGenOres implements IWorldGenerator {
+	@Override
+	public void generate(Random random, int chunkX, int chunkZ, World world, IChunkProvider chunkGenerator, IChunkProvider chunkProvider) {
+		// TODO Auto-generated method stub
+		
+	}
+}