+++ src/main/java/com/sammccreery/cherry/registry/CherryBlocks.java
@@ -1,19 +1,21 @@
 package com.sammccreery.cherry.registry;
 
-import com.sammccreery.cherry.util.ResourceName;
-import com.sammccreery.cherry.util.ResourceName.Format;
+import com.sammccreery.cherry.util.UniversalName;
+import com.sammccreery.cherry.util.UniversalName.Format;
 
 import cpw.mods.fml.common.registry.GameRegistry;
 import net.minecraft.block.Block;
 
-public final class CherryBlocks {
-	private CherryBlocks() {}
-
-	public static Block registerBlock(Block block, ResourceName name) {
+public final class CherryBlocks extends Registry<Block> {
+	@Override
+	public Block registerLocal(Block block, UniversalName name) {
 		block.setBlockName(name.format(false, Format.HEADLESS));
 		block.setBlockTextureName(name.format(true, Format.SNAKE));
 		GameRegistry.registerBlock(block, name.format(false, Format.SNAKE));
 
 		return block;
 	}
+
+	@Override
+	public void init() {}
 }