+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Unreleased]
 ### Added
+- Generic global registry and naming system
 - Base resources and tools from Tools Done Right